Ensured percentages on stats tables are ints
git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/mozootle@7749 54714841-351b-0410-a198-e36a94b762f5

+++ pagelayout.py
@@ -303,7 +303,7 @@ class PootleNavPage(PootlePage):
     for key in wanted:
       percentkey = key + "percentage"
       wordkey = key + "sourcewords"
-      gotten[percentkey] = gotten[wordkey]*100/max(gotten["totalsourcewords"], 1)
+      gotten[percentkey] = int(gotten[wordkey]*100/max(gotten["totalsourcewords"], 1))
 
     for key in gotten:
       if key.find("check-") == 0:
